// front-end/src/models/stats.model.ts
export interface Stats {
  quizCompletionRate: number;
  improvementCorrectRate: number;
  averageResponseTime: number;
  totalQuestions: number;
  avgImagesPerQuestion: number;
  totalQuizzes: number;
  positiveResponseTrend: string;
  reactionTimeTrend: string;
  helpUsageTrend: string;
  chartData: any;
  chartOptions: any;
  evolutionTimeData: any;
  evolutionImagesData: any;
  quizTypePieData: any;
}